+++ termal_core/src/raw/events/status.rs
@@ -1,8 +1,9 @@
 use super::TermAttr;
 
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone)]
 pub enum Status {
     Attributes(TermAttr),
     Ok,
     CursorPosition { x: usize, y: usize },
+    TerminalName(String),
 }